requirements.txt
flask>=2.0.0
sqlalchemy>=1.4.0
pymysql>=1.0.0
streamlit>=1.0.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.0
plotly>=5.0.0
folium>=0.12.0
bcrypt>=3.2.0
pyjwt>=2.1.0
marshmallow>=3.14.0
requests>=2.26.0
cryptography>=3.4.0
python-dotenv>=0.19.0  # 用于环境变量管理 